add logout button functionality to menu
