+++ Agenda-client/src/App.tsx
@@ -6,9 +6,10 @@ import { Navbar } from './components/Navbar/Navbar';
 import { LogIn } from './components/LoginPage/LogIn';
 import Token from './components/Token/Token';
 import { createContext, useState, ReactNode } from 'react';
-
-export const DashboardContext = createContext<any>(null)
 import { LectureProvider } from './context/LectureContext';
+
+export const DashboardContext = createContext<any>(null);
+export const AuthContext = createContext<any>(null);
 
 function App() {
   const [dashboardState, setDashboardState] = useState<DashboardState>();
@@ -16,44 +17,48 @@ function App() {
 
   return (
     <>
-    <DashboardContext.Provider value={{dashboardState, setDashboardState}}>
-      <BrowserRouter>
-        <Navbar />
-        <LectureProvider>
-          <Routes>
-            <Route
-              path="/"
-              element={<LogIn setIsAuthenticated={setIsAuthenticated} />}
-            />
-            <Route path="/token/:token" element={<Token />} />
-            <Route
-              path="/dashboard"
-              element={
-                <ProtectedRoute isAuthenticated={isAuthenticated}>
-                  <Dashboard />
-                </ProtectedRoute>
-              }
-            />
-            <Route
-              path="/lecture/:lectureId"
-              element={
-                <ProtectedRoute isAuthenticated={isAuthenticated}>
-                  <LectureComponent />
-                </ProtectedRoute>
-              }
-            />
-            <Route
-              path="/syllabus"
-              element={
-                <ProtectedRoute isAuthenticated={isAuthenticated}>
-                  <SyllabusPage />
-                </ProtectedRoute>
-              }
-            />
-          </Routes>
-        </LectureProvider>
-      </BrowserRouter>
-      </DashboardContext.Provider>
+      <AuthContext.Provider value={{ isAuthenticated, setIsAuthenticated }}>
+        <DashboardContext.Provider
+          value={{ dashboardState, setDashboardState }}
+        >
+          <BrowserRouter>
+            <Navbar isAuthenticated={isAuthenticated} />
+            <LectureProvider>
+              <Routes>
+                <Route
+                  path="/"
+                  element={<LogIn setIsAuthenticated={setIsAuthenticated} />}
+                />
+                <Route path="/token/:token" element={<Token />} />
+                <Route
+                  path="/dashboard"
+                  element={
+                    <ProtectedRoute isAuthenticated={isAuthenticated}>
+                      <Dashboard />
+                    </ProtectedRoute>
+                  }
+                />
+                <Route
+                  path="/lecture/:lectureId"
+                  element={
+                    <ProtectedRoute isAuthenticated={isAuthenticated}>
+                      <LectureComponent />
+                    </ProtectedRoute>
+                  }
+                />
+                <Route
+                  path="/syllabus"
+                  element={
+                    <ProtectedRoute isAuthenticated={isAuthenticated}>
+                      <SyllabusPage />
+                    </ProtectedRoute>
+                  }
+                />
+              </Routes>
+            </LectureProvider>
+          </BrowserRouter>
+        </DashboardContext.Provider>
+      </AuthContext.Provider>
     </>
   );
 }
