+++ Agenda-client/src/components/Navbar/Navbar.tsx
@@ -4,7 +4,7 @@ import { IoIosMenu, IoIosClose } from 'react-icons/io';
 import Menu from '../Menu/Menu';
 import logo from '../../assets/LOGO.png';
 
-export const Navbar: FC = () => {
+export const Navbar: FC<{isAuthenticated:boolean}> = ({isAuthenticated}) => {
   const [showMenu, setShowMenu] = useState(false);
 
   function handleMenuClick() {
@@ -24,15 +24,20 @@ export const Navbar: FC = () => {
             data-ll-status="loaded"
           ></img>
         </Link>
-        <div
-          onClick={handleMenuClick}
-          className="cursor-pointer w-[50px] h-full hover:bg-black hover:text-white flex justify-center"
-        >
-          {showMenu ? <IoIosClose></IoIosClose> : <IoIosMenu></IoIosMenu>}
-        </div>
-        {showMenu && <Menu></Menu>}
-        {/*    TODO: if there will be  a sidebar, this will be a button to open the sidebar
-        <div>open the sidebar to set widgets</div> */}
+        {
+        isAuthenticated &&
+        <>
+          <div
+            onClick={handleMenuClick}
+            className="cursor-pointer w-[50px] h-full p-4 -mt-1 rounded-md text-cp-dark-blue bg-cp-light-blue hover:bg-cp-dark-blue hover:text-white flex justify-center"
+          >
+            {showMenu ? <IoIosClose></IoIosClose> : <IoIosMenu></IoIosMenu>}
+          </div>
+          {showMenu && <Menu></Menu>}
+          {/*    TODO: if there will be  a sidebar, this will be a button to open the sidebar
+          <div>open the sidebar to set widgets</div> */}
+        </>
+      }
       </div>
     </div>
   );
