+++ Agenda-client/src/components/userProfile/UserProfile.tsx
@@ -1,15 +1,13 @@
 import { FC, useEffect, useState } from 'react';
-import logoutIcon from '../../assets/fi-rs-sign-out.svg';
 import IUser from '../../utils/types';
 import userData from './data/userData.json';
 import { createInitialsAvatar } from '../../utils/createInitialsAvatar';
-import { useNavigate } from 'react-router-dom';
+import LogoutButton from './LogoutButton';
 
 export const UserProfile: FC = () => {
   const [user, setUser] = useState<IUser>();
   const [avatar, setAvatar] = useState<string>('');
 
-  const navigate = useNavigate();
 
   useEffect(() => {
     const user: IUser = {
@@ -25,11 +23,6 @@ export const UserProfile: FC = () => {
     if (!user.avatar) setAvatar(createInitialsAvatar(user));
     else setAvatar(user.avatar);
   }, []);
-
-  const handleLogout = () => {
-    localStorage.removeItem('token');
-    navigate('/');
-  };
 
   return user ? (
     <div className="profile flex justify-between w-full h-full p-2 overflow-hidden">
@@ -51,13 +44,7 @@ export const UserProfile: FC = () => {
         </div>
       </div>
       <div className="flex flex-end grow-0 shrink-0">
-        <button
-          data-testid="logoutButton"
-          className="text-white pt-2 pl-4 pr-4 pb-2 h-fit rounded-[0.5rem] bg-cp-blue hover:bg-cp-middle-blue active:scale-90 shadow-md active:shadow-inner"
-          onClick={handleLogout}
-        >
-          <img src={logoutIcon} width={16} height={16} />
-        </button>
+        <LogoutButton></LogoutButton>
       </div>
     </div>
   ) : (
